Change the project name
The name "calabash" is somewhat misleading, as the project do not mimic _bash_
but rather text stream manipulation in Un*x-like shells.

As it tries to do so in _pure_ python (there is better alternatives if one want
to use the shell commands themselves, like Plumbum), the new name reads:
py[thon][Uni]xshell, which sounds a bit like "pixel", which remains the atomic
nature of shell commands.

Also, this new name is not already google-taken.